refactor: 캐싱 비즈니스 로직과 기존 DB 비즈니스 로직을 두 클래스로 병확하게 분리
