+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/tuningreport/service/TuningReportService.java
@@ -1,7 +1,6 @@
 package com.hertz.hertz_be.domain.tuningreport.service;
 
 import com.hertz.hertz_be.domain.tuningreport.dto.response.TuningReportListResponse;
-import com.hertz.hertz_be.domain.tuningreport.entity.TuningReport;
 import com.hertz.hertz_be.domain.tuningreport.entity.TuningReportUserReaction;
 import com.hertz.hertz_be.domain.tuningreport.entity.enums.ReactionType;
 import com.hertz.hertz_be.domain.tuningreport.entity.enums.TuningReportSortType;
@@ -10,11 +9,9 @@ import com.hertz.hertz_be.domain.tuningreport.repository.TuningReportRepository;
 import com.hertz.hertz_be.domain.tuningreport.repository.TuningReportUserReactionRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
@@ -26,22 +23,20 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class TuningReportService {
 
-    private final TuningReportRepository tuningReportRepository;
-    private final TuningReportUserReactionRepository tuningReportUserReactionRepository;
+    private final TuningReportTransactionalService transactionalService;
     private final TuningReportCacheManager cacheManager;
     private final RedisTemplate<String, String> redisTemplate;
+    private final TuningReportRepository tuningReportRepository;
+    private final TuningReportUserReactionRepository tuningReportUserReactionRepository;
 
-    @Transactional(readOnly = true)
     public TuningReportListResponse getReportList(Long userId, int page, int size, TuningReportSortType sort) {
         if (isCacheApplicable(page, size, sort)) {
             List<TuningReportListResponse.ReportItem> cachedItems = loadOrCacheReports(page, size, sort);
-
             List<TuningReportListResponse.ReportItem> enriched = enrichWithUserReactions(cachedItems, userId);
-
             return createResponse(enriched, page, size);
         }
 
-        return fetchDirectlyFromDB(userId, page, size, sort);
+        return transactionalService.fetchDirectlyFromDB(userId, page, size, sort);
     }
 
     private boolean isCacheApplicable(int page, int size, TuningReportSortType sort) {
@@ -53,25 +48,14 @@ public class TuningReportService {
         if (items == null) {
             log.info("게시글 리스트 반환 시 캐싱 hit ⚠️");
             var pageReq = PageRequest.of(page, size);
-            var reports = sort.fetch(pageReq, tuningReportRepository);
+            var reports = sort.fetch(pageReq, transactionalService.getTuningReportRepository());
             items = reports.stream()
-                    .map(this::toReportItemWithoutReactions)
+                    .map(transactionalService::toReportItemWithoutReactions)
                     .collect(Collectors.toList());
             cacheManager.cacheReportList(items);
         }
         log.info("캐싱 되기 위해 load된 item 수: {}", items.size());
         return items;
-    }
-
-    private TuningReportListResponse.ReportItem toReportItemWithoutReactions(TuningReport report) {
-        return new TuningReportListResponse.ReportItem(
-                report.getCreatedAt(), report.getId(), report.getTitle(), report.getContent(),
-                new TuningReportListResponse.Reactions (
-                        report.getReactionCelebrate(), report.getReactionThumbsUp(),
-                        report.getReactionLaugh(), report.getReactionEyes(), report.getReactionHeart()
-                ),
-                null
-        );
     }
 
     private List<TuningReportListResponse.ReportItem> enrichWithUserReactions(List<TuningReportListResponse.ReportItem> items, Long userId) {
@@ -80,11 +64,10 @@ public class TuningReportService {
                 .toList();
 
         List<TuningReportUserReaction> dbList =
-                tuningReportUserReactionRepository.findAllByUserIdAndReportIdIn(userId, reportIds);
+                transactionalService.getTuningReportUserReactionRepository().findAllByUserIdAndReportIdIn(userId, reportIds);
 
         log.info("💡 사용자 {}에 대해 DB에서 조회된 반응 수: {}", userId, dbList.size());
 
-        // 캐싱 먼저 수행
         dbList.forEach(reaction ->
                 cacheManager.setUserReaction(
                         reaction.getReport().getId(),
@@ -94,14 +77,12 @@ public class TuningReportService {
                 )
         );
 
-        // 사용자별 반응 목록 맵핑
         Map<Long, Set<ReactionType>> userReactionMap = dbList.stream()
                 .collect(Collectors.groupingBy(
                         r -> r.getReport().getId(),
                         Collectors.mapping(TuningReportUserReaction::getReactionType, Collectors.toSet())
                 ));
 
-        // 최종 응답 객체 변환
         return items.stream()
                 .map(item -> {
                     Set<ReactionType> reactions = userReactionMap.getOrDefault(item.getReportId(), Set.of());
@@ -127,56 +108,4 @@ public class TuningReportService {
         boolean isLast = items.size() < size;
         return new TuningReportListResponse(items, page, size, isLast);
     }
-
-    private TuningReportListResponse fetchDirectlyFromDB(Long userId, int page, int size, TuningReportSortType sort) {
-        PageRequest pageRequest = PageRequest.of(page, size);
-        Page<TuningReport> reports = sort.fetch(pageRequest, tuningReportRepository);
-
-        List<Long> reportIds = reports.stream().map(TuningReport::getId).toList();
-        List<TuningReportUserReaction> userReactions =
-                tuningReportUserReactionRepository.findAllByUserIdAndReportIdIn(userId, reportIds);
-
-        Map<Long, Set<ReactionType>> userReactionMap = userReactions.stream()
-                .collect(Collectors.groupingBy(
-                        r -> r.getReport().getId(),
-                        Collectors.mapping(TuningReportUserReaction::getReactionType, Collectors.toSet())
-                ));
-
-        List<TuningReportListResponse.ReportItem> reportItems = reports.stream()
-                .map(report -> toReportItemWithReactions(report, userReactionMap.getOrDefault(report.getId(), Set.of())))
-                .collect(Collectors.toList());
-
-        log.info("DB에서 바로 반환된 item 수: {}", reportItems.size());
-        return new TuningReportListResponse(
-                reportItems,
-                reports.getNumber(),
-                reports.getSize(),
-                reports.isLast()
-        );
-    }
-
-    private TuningReportListResponse.ReportItem toReportItemWithReactions(TuningReport report, Set<ReactionType> myReactions) {
-        return new TuningReportListResponse.ReportItem(
-                report.getCreatedAt(),
-                report.getId(),
-                report.getTitle(),
-                report.getContent(),
-                new TuningReportListResponse.Reactions(
-                        report.getReactionCelebrate(),
-                        report.getReactionThumbsUp(),
-                        report.getReactionLaugh(),
-                        report.getReactionEyes(),
-                        report.getReactionHeart()
-                ),
-                new TuningReportListResponse.MyReactions(
-                        myReactions.contains(ReactionType.CELEBRATE),
-                        myReactions.contains(ReactionType.THUMBS_UP),
-                        myReactions.contains(ReactionType.LAUGH),
-                        myReactions.contains(ReactionType.EYES),
-                        myReactions.contains(ReactionType.HEART)
-                )
-        );
-    }
-
-
 }
